refactor(supports): import getSettings from the state module, not the barrel
The Settings barrel re-exports SupportSidebar, so every engine file that
reached for getSettings pulled the whole settings UI into its import graph —
and with it, once the sidebar is translated, a Lingui macro the test runner
cannot transform.
